fix: skip slots with active reading `Ref`s in `push_ref` (#81)
`Core::push_ref` can go into an (almost infinite) spin loop waiting for
a `Ref` (created in `pop_ref`) to this slot to be dropped. This
behaviour can lead to writing being blocked unless all refs are dropped,
even if we have free space in the buffer. 

In this PR I've added a mechanism to skip such slots (by updating their
`state`) and attempt to write into them on the next lap. 

Fixes #83
Closes #80